# Summative-2-Olaniran-Obaloluwa/src/main/java/com/company/bookstore/controller/AuthorController.java
package com.company.bookstore.controller;

import com.company.bookstore.model.Author;
import com.company.bookstore.repository.AuthorRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
public class AuthorController {

    @Autowired
    AuthorRepository authorRepo;

    @PostMapping("/author")
    @ResponseStatus(HttpStatus.CREATED)
    public Author addAuthor(@RequestBody Author author) {
        return authorRepo.save(author);
    }


    @GetMapping("/author/{authorId}")
    public Author getAuthorById(@PathVariable int authorId) {

        Optional<Author> returnVal = authorRepo.findById(authorId);
        if (returnVal.isPresent()) {
            return returnVal.get();
        } else {
            return null;
        }
    }

    @GetMapping("/author")
    public List<Author> getAllAuthors() {
        return authorRepo.findAll();
    }

    @PutMapping("/author")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void updateAuthor(@RequestBody Author author) {
        authorRepo.save(author);
    }

    @DeleteMapping("/author/{authorId}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void deleteAuthor(@PathVariable int authorId) {
        authorRepo.deleteById(authorId);
    }

}