# Summative-2-Olaniran-Obaloluwa/src/main/java/com/company/bookstore/controller/BookController.java
package com.company.bookstore.controller;

import com.company.bookstore.model.Book;
import com.company.bookstore.repository.BookRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.HttpStatus;
import org.springframework.web.bind.annotation.*;

import java.util.List;
import java.util.Optional;

@RestController
public class BookController {

    @Autowired
    BookRepository bookRepo;

    @PostMapping("/books")
    @ResponseStatus(HttpStatus.CREATED)
    public Book addBook(@RequestBody Book book) {
        return bookRepo.save(book);
    }

    @GetMapping("/book/{bookId}")
    public Book getBookById(@PathVariable int bookId) {

        Optional<Book> returnVal = bookRepo.findById(bookId);
        if (returnVal.isPresent()) {
            return returnVal.get();
        } else {
            return null;
        }
    }

    @GetMapping("/book/author/{authorId}")
    public Book searchBookByAuthorId(@PathVariable int authorId) {
        Optional<Book> returnVal = bookRepo.findById(authorId);
        if (returnVal.isPresent()) {
            return returnVal.get();
        } else {
            return null;
        }
    }

    @GetMapping("/book")
    public List<Book> getBooks() {
        return bookRepo.findAll();
    }


    @PutMapping("/book")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void updateBook(@RequestBody Book book) {
        bookRepo.save(book);
    }

    @DeleteMapping("/book/{bookId}")
    @ResponseStatus(HttpStatus.NO_CONTENT)
    public void deleteBook(@PathVariable int bookId) {
        bookRepo.deleteById(bookId);
    }
}